+++ src/Luxor.jl
@@ -65,7 +65,7 @@ include("noise.jl")
 include("deprecations.jl")
 include("randompoints.jl")
 include("Path.jl")
-# include("latex.jl")
+include("precompile.jl")
 # include("play.jl") # will require MiniFB
 # include("shapefile.jl") # don't load unless you've loaded Shapefile.jl