+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- hexspiral() counting from 1 not 0 (or 2)
+
 ### Removed
 
 ### Deprecated